requirements.txt
metar
requests
pillow
python-dateutil
tzlocal
geocoder